requirements.txt
streamlit>=1.35.0
PyPDF2>=3.0.1
google-generativeai>=0.8.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.24.0
torch>=2.0.0